requirements.txt
absl-py==0.6.0
asgiref==3.4.1
astor==0.7.1
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
Django==3.2.19
djangorestframework==3.14.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3
networkx==2.2
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
pkg-resources==0.0.0
protobuf==3.6.1
psycopg2==2.9.6
pyparsing==2.2.2
python-dateutil==2.7.4
pytz==2023.3
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
sqlparse==0.4.4
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
tfmpl==1.0.2
toolz==0.9.0
typing-extensions==4.1.1
Werkzeug==0.14.1
yapf==0.24.0
